+++ src/harlequin/tui/components/schema_viewer.py
@@ -1,3 +1,4 @@
+from dataclasses import dataclass
 from typing import Generic, List, Set, Tuple, Union
 
 from duckdb import DuckDBPyConnection
@@ -12,7 +13,13 @@ from harlequin.duck_ops import Catalog
 from harlequin.tui.utils import short_type
 
 
-class SchemaViewer(Tree[str]):
+@dataclass
+class CatalogItem:
+    qualified_identifier: str
+    query_name: str
+
+
+class SchemaViewer(Tree[CatalogItem]):
     class NodeSubmitted(Generic[EventTreeDataType], Message):
         def __init__(self, node: TreeNode[EventTreeDataType]) -> None:
             self.node: TreeNode[EventTreeDataType] = node
@@ -40,7 +47,7 @@ class SchemaViewer(Tree[str]):
         label: TextType,
         connection: DuckDBPyConnection,
         type_color: str = "#888888",
-        data: Union[str, None] = None,
+        data: Union[CatalogItem, None] = None,
         name: Union[str, None] = None,
         id: Union[str, None] = None,  # noqa: A002
         classes: Union[str, None] = None,
@@ -52,7 +59,7 @@ class SchemaViewer(Tree[str]):
         super().__init__(
             label, data, name=name, id=id, classes=classes, disabled=disabled
         )
-        self.double_click = False
+        self.double_click: Union[int, None] = None
 
     def on_mount(self) -> None:
         self.border_title = self.label
@@ -73,14 +80,14 @@ class SchemaViewer(Tree[str]):
                 database_identifier = f'"{database[0]}"'
                 database_node = self.root.add(
                     database[0],
-                    data=database_identifier,
+                    data=CatalogItem(database_identifier, database_identifier),
                     expand=database_identifier in expanded_nodes,
                 )
                 for schema in database[1]:
                     schema_identifier = f'{database_identifier}."{schema[0]}"'
                     schema_node = database_node.add(
                         schema[0],
-                        data=schema_identifier,
+                        data=CatalogItem(schema_identifier, schema_identifier),
                         expand=schema_identifier in expanded_nodes,
                     )
                     for table in schema[1]:
@@ -88,24 +95,27 @@ class SchemaViewer(Tree[str]):
                         table_identifier = f'{schema_identifier}."{table[0]}"'
                         table_node = schema_node.add(
                             f"{table[0]} [{self.type_color}]{short_table_type}[/]",
-                            data=table_identifier,
+                            data=CatalogItem(table_identifier, table_identifier),
                             expand=table_identifier in expanded_nodes,
                         )
                         for col in table[2]:
-                            col_identifier = f'{table_identifier}."{col[0]}"'
+                            col_name = f'"{col[0]}"'
+                            col_identifier = f"{table_identifier}.{col_name}"
                             table_node.add_leaf(
                                 f"{col[0]} [{self.type_color}]{short_type(col[1])}[/]",
-                                data=col_identifier,
+                                data=CatalogItem(col_identifier, col_name),
                             )
 
     @classmethod
-    def get_node_states(cls, node: TreeNode[str]) -> Tuple[List[str], Union[str, None]]:
+    def get_node_states(
+        cls, node: TreeNode[CatalogItem]
+    ) -> Tuple[List[str], Union[str, None]]:
         expanded_nodes = []
         selected_node = None
         if node.is_expanded and node.data is not None:
-            expanded_nodes.append(node.data)
+            expanded_nodes.append(node.data.qualified_identifier)
         if node._selected and node.data is not None:
-            selected_node = node.data
+            selected_node = node.data.qualified_identifier
         for child in node.children:
             expanded_children, selected_child = cls.get_node_states(child)
             expanded_nodes.extend(expanded_children)
@@ -113,7 +123,7 @@ class SchemaViewer(Tree[str]):
         return expanded_nodes, selected_node
 
     def _clear_double_click(self) -> None:
-        self.double_click = False
+        self.double_click = None
 
     async def on_click(self, event: Click) -> None:
         """
@@ -123,16 +133,19 @@ class SchemaViewer(Tree[str]):
         we collapse them, since the _on_click event will go after this and toggle
         their state.
         """
-        if self.double_click:
-            meta = event.style.meta
-            cursor_line = meta.get("line", None)
-            if cursor_line is not None:
-                node = self.get_node_at_line(cursor_line)
-                if node is not None:
-                    self.post_message(self.NodeSubmitted(node=node))
-                    node.collapse()
+        meta = event.style.meta
+        click_line: Union[int, None] = meta.get("line", None)
+        if (
+            self.double_click is not None
+            and click_line is not None
+            and self.double_click == click_line
+        ):
+            node = self.get_node_at_line(click_line)
+            if node is not None:
+                self.post_message(self.NodeSubmitted(node=node))
+                node.collapse()
         else:
-            self.double_click = True
+            self.double_click = click_line
             self.set_timer(
                 delay=0.5, callback=self._clear_double_click, name="double_click_timer"
             )