+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Features
 
--   Double-click or press <kbd>ctrl+enter</kbd> on an item in the data catalog to insert the name in the query editor. ([#194](https://github.com/tconbeer/harlequin/issues/194)).
+-   Double-click or press <kbd>ctrl+enter</kbd> on an item in the data catalog to insert the name in the query editor ([#194](https://github.com/tconbeer/harlequin/issues/194)).
 
 ### Bug Fixes and Minor Updates